+++ include/Data_writer_buffer.h
@@ -29,7 +29,7 @@ public:
   ~Data_writer_buffer();
 
 private:
-  INT64 do_put_bytes(INT64 nBytes, char *buff);
+  size_t do_put_bytes(size_t nBytes, char *buff);
 
 
   // The input buffer
